Solve type errors in log interface.

+++ source/metacall/include/metacall/metacall_log.h
@@ -118,9 +118,9 @@ struct metacall_log_custom_va_list_type
 struct metacall_log_custom_type
 {
 	void *context;
-	size_t (*format_size)(void *, const char *, size_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
-	size_t (*format_serialize)(void *, void *, const size_t, const char *, size_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
-	size_t (*format_deserialize)(void *, const void *, const size_t, const char *, size_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
+	size_t (*format_size)(void *, const char *, uint64_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
+	size_t (*format_serialize)(void *, void *, const size_t, const char *, uint64_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
+	size_t (*format_deserialize)(void *, const void *, const size_t, const char *, uint64_t, size_t, const char *, const char *, const char *, const char *, metacall_log_custom_va_list);
 	int (*stream_write)(void *, const char *, const size_t);
 	int (*stream_flush)(void *);
 };
